Moved separate core collections into their own files in their own package
